dbchat-app/frontend/dbchat-frontend/pyproject.toml
[tool.poetry]
name = "dbchat-frontend"
version = "0.1.0"
description = ""
authors = ["Florencia Froebel <ff4348@berkeley.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.32.2"
requests = "^2.31.0"
langchain-openai = "^0.1.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
